+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Implement details view for match profiles (UIDATIMP-140)
 * Save the new match profile (UIDATIMP-139)
 * Create edit match profile form (UIDATIMP-143)
+* Implement saving of edited match profile (UIDATIMP-182)
 
 ## [1.2.1](https://github.com/folio-org/ui-data-import/tree/v1.2.1) (2019-05-14)
 * Downgrade okapi interface version in order to have data-import-converter-storage API endpoints work (UIDATIMP-195)